+++ client/src/main.js
@@ -13,6 +13,12 @@ monkeypatch(Craft.AssetIndex, 'init', function () {
   const replaceButton = new Button('Replace')
 
   const $uploadButton = this.$uploadButton
+
+  // If there's no upload button, there should be no embed button
+  if (!$uploadButton) {
+    return
+  }
+
   const inHeader = $uploadButton.closest('#header').length > 0
   const inModal = $uploadButton.closest('.modal').length > 0
 
@@ -49,6 +55,12 @@ monkeypatch(Craft.AssetIndex, 'init', function () {
   replaceButton.hide()
 
   const getFolderId = () => {
+    if (typeof this.sourcePath !== 'undefined') {
+      // Craft 3.8 subfolder compatibility
+      const currentFolder = this.sourcePath[this.sourcePath.length - 1]
+      return currentFolder.folderId
+    }
+
     const split = this.sourceKey.split(':')
 
     if (split[split.length - 1]) {